% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datawrapper_auth.R
\name{datawrapper_auth}
\alias{datawrapper_auth}
\title{Set Datawrapper-API-Key to system environment}
\usage{
datawrapper_auth(api_key)
}
\arguments{
\item{api_key}{Required. A character string, containing the API-Key.}
}
\value{
A Message in the command line.
}
\description{
Adds Key to Environment to be available on Startup.
}
\note{
This is a very simple function that adds the API-key to the .Renviron-file in the user's home folder. If a key already exists and the user requests to, it will get replaced.
}
\examples{

\dontrun{datawrapper_auth(api_key = "1234ABC")}
}
\author{
Benedict Witzenberger
}
