% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatawRappr-package.R
\docType{package}
\name{DatawRappr-package}
\alias{DatawRappr}
\alias{DatawRappr-package}
\title{Provides Access to Datawrapper's API}
\description{
Connecting R to Datawrapper, allowing the creation and edits of charts via Datawrapper's API.
It makes heavy use of the httr-library and tries to manage the API-token for the user in the .Renviron-file.
\if{html}{\figure{logo.png}{options: align="right" style="width:125px; height:145px; padding-top:16px"}}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/munichrocker/DatawRappr}
  \item Report bugs at \url{https://github.com/munichrocker/DatawRappr/issues}
}

}
\author{
\strong{Maintainer}: Benedict Witzenberger \email{b.witzenberger@googlemail.com}

Authors:
\itemize{
  \item Bob Rudis \email{bob@@rud.is}
}

}
\keyword{internal}
