% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dw_copy_chart.R
\name{dw_copy_chart}
\alias{dw_copy_chart}
\title{Makes a copy of an existing Datawrapper chart}
\usage{
dw_copy_chart(copy_from = "", api_key = "environment")
}
\arguments{
\item{copy_from}{Required. Chart_id from an existing chart which serves as the template.}

\item{api_key}{Required. A Datawrapper-API-key as character string. Defaults to "environment" - tries to automatically retrieve the key that's stored in the .Reviron-file by \code{\link{datawrapper_auth}}.}
}
\value{
It prints the new chart's id and returns a S3-structure of type \strong{dw_chart} with the elements from the Datawrapper-API, the same as in \code{\link{dw_retrieve_chart_metadata}} and \code{\link{dw_create_chart}}.
}
\description{
\lifecycle{maturing}
Creates a new Datawrapper chart object from an existing chart. This function starts the chart-making process like \code{\link{dw_create_chart}}.
}
\examples{

\dontrun{
dw_copy_chart(copy_from = "ABCDE")
# uses api-key from environment - if set.
}

\dontrun{
dw_copy_chart(copy_from = "ABCDE", api_key = "1234ABCD")
} # uses the specified key

}
\author{
Benedict Witzenberger
}
