% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dw_create_chart.R
\name{dw_create_chart}
\alias{dw_create_chart}
\title{Creates a new Datawrapper chart}
\usage{
dw_create_chart(api_key = "environment")
}
\arguments{
\item{api_key}{Optional. A Datawrapper-API-key as character string. Defaults to "environment" - tries to automatically retrieve the key that's stored in the .Reviron-file by \code{\link{datawrapper_auth}}.}
}
\value{
A list with the elements from the Datawrapper-API, the same as in dw_retrieve_chart_metadata()
}
\description{
Creates and returns the metadata of the new Datawrapper chart.
}
\note{
This function retrieves all metadata about the new chart that's stored by Datawrapper. The new chart will by default be created without a title and with the type d3-lines.
}
\examples{

dw_create_chart() # uses the preset key in the .Renviron-file

dw_create_chart(api_key = "1234ABCD") # uses the specified key
}
\author{
Benedict Witzenberger
}
