% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dw_create_folder.R
\name{dw_create_folder}
\alias{dw_create_folder}
\title{Creates a new folder}
\usage{
dw_create_folder(
  name = "",
  organization_id = NULL,
  parent_id = NULL,
  api_key = "environment"
)
}
\arguments{
\item{name}{Required. Specify the name for the newly created folder.}

\item{organization_id}{Optional. ID of organization if creating a folder for a organization that you're working with. Defaults to user.}

\item{parent_id}{Optional. ID of parent folder if creating a subfolder.}

\item{api_key}{Optional. A Datawrapper-API-key as character string. Defaults to "environment" - tries to automatically retrieve the key that's stored in the .Reviron-file by \code{\link{datawrapper_auth}}.}
}
\value{
Returns a success message
}
\description{
\lifecycle{maturing}
Creates a new folder
}
\note{
This function builds a body for a API-call to the Datawrapper-API, and creates a new folder for a user or a team.
}
\examples{

\dontrun{

dw_create_folder(name = "New Folder")

} # uses the preset API-key in the .Renviron-file

}
\author{
Benedict Witzenberger
}
