% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dw_create_locator_map.R
\name{dw_create_locator_map}
\alias{dw_create_locator_map}
\title{Creates a new Datawrapper Locator Map}
\usage{
dw_create_locator_map(
  markers = list(),
  options = list(),
  title = NULL,
  folderId = NULL,
  api_key = "environment"
)
}
\arguments{
\item{markers}{Optional. A list of markers with details such as latitude, longitude, and label.}

\item{options}{Optional. A list of map options (e.g., basemap, zoom level).}

\item{title}{Optional. Will set the locator map's title on creation}

\item{folderId}{Optional. Creates chart in specified folder.}

\item{api_key}{API key for Datawrapper. If not provided, the function will use the key set with `dw_set_api_key()`.}
}
\value{
It prints the new chart's id and returns a S3-structure of type \strong{dw_chart} with the elements from the Datawrapper-API, the same as in \code{\link{dw_retrieve_chart_metadata}}.
}
\description{
\lifecycle{experimental}
This function creates a new Locator Map in Datawrapper using the Datawrapper API. This function starts the map-making process
}
\note{
If not specified, the new chart will by default be created without a title.
}
\author{
Duc Quang Nguyen
}
