% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dw_data_from_chart.R
\name{dw_data_from_chart}
\alias{dw_data_from_chart}
\title{Download the data in a Datawrapper chart}
\usage{
dw_data_from_chart(chart_id, api_key = "environment")
}
\arguments{
\item{chart_id}{Required. A Datawrapper-chart-id as character string, usually a five character combination of digits and letters, e.g. "aBcDe". Or a \emph{dw_chart}-object. You need to have access to the chart (e.g. it was created by yourself or in one of your teams).}

\item{api_key}{Optional. A Datawrapper-API-key as character string. Defaults to "environment" - tries to automatically retrieve the key that's stored in the .Reviron-file by \code{\link{datawrapper_auth}}.}
}
\value{
A data.frame.
}
\description{
\lifecycle{maturing}
Downloads the data from an existing Datawrapper chart, that ypu have access to. This is the mirrored function to \code{\link{dw_data_to_chart}}.
}
\note{
This function downloads a R-dataframe from a Datawrapper chart.
}
\examples{

\dontrun{dw_data_from_chart(df, "aBcDE")} # uses the preset key in the .Renviron-file

\dontrun{df <- dw_data_from_chart(df, chart_id = "a1B2Cd", api_key = "1234ABCD")} # uses the specified key

}
\author{
Benedict Witzenberger
}
