% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dw_delete_chart.R
\name{dw_delete_chart}
\alias{dw_delete_chart}
\title{Delete a Datawrapper chart}
\usage{
dw_delete_chart(chart_id, api_key = "environment")
}
\arguments{
\item{chart_id}{Required. A Datawrapper-chart-id as character string, usually a five character combination of digits and letters, e.g. "aBcDe".}

\item{api_key}{Optional. A Datawrapper-API-key as character string. Defaults to "environment" - tries to automatically retrieve the key that's stored in the .Reviron-file.}
}
\value{
A message that specifies, if the deletion was successfull.
}
\description{
Deletes a chart on Datawrapper.
}
\note{
This function deletes a chart in Datawrapper.
}
\examples{

dw_delete_chart("aBcDE") # uses the preset key in the .Renviron-file

dw_delete_chart(chart_id = "a1B2Cd", api_key = "1234ABCD") # uses the specified key

}
\author{
Benedict Witzenberger
}
