% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dw_edit_chart.R
\name{dw_edit_chart}
\alias{dw_edit_chart}
\title{Edits a existing Datawrapper chart}
\usage{
dw_edit_chart(
  chart_id,
  api_key = "environment",
  title = "",
  intro = "",
  annotate = "",
  type = "",
  source_name = "",
  source_url = "",
  data = list(),
  visualize = list(),
  describe = list(),
  publish = list()
)
}
\arguments{
\item{chart_id}{Required. A Datawrapper-chart-id as character string, usually a five character combination of digits and letters, e.g. "aBcDe". Or a \strong{dw_chart}-object.}

\item{api_key}{Optional. A Datawrapper-API-key as character string. Defaults to "environment" - tries to automatically retrieve the key that's stored in the .Reviron-file by \code{\link{datawrapper_auth}}.}

\item{title}{Optional. Adds a title to the plot.}

\item{intro}{Optional. Adds an intro below the title.}

\item{annotate}{Optional. Adds a annotation below the plot.}

\item{type}{Optional. Changes the type of the chart. See \href{https://developer.datawrapper.de/docs/chart-types-2}{the documentation} for the different types.}

\item{source_name}{Optional. Adds a source name to the plot.}

\item{source_url}{Optional. Adds a URL to the source name (displayed only, if source name specified). Include http(s):// before URL.}

\item{data}{Optional. A list. Add separate arguments for the data. See \href{https://developer.datawrapper.de/docs/chart-properties-1}{the documentation} for details.}

\item{visualize}{Optional. A list. Add separate arguments for the visualization. See \href{https://developer.datawrapper.de/docs/chart-properties-1}{the documentation} for details.}

\item{describe}{Optional. A list. Add separate arguments for the description. See \href{https://developer.datawrapper.de/docs/chart-properties-1}{the documentation} for details.}

\item{publish}{Optional. A list. Add separate arguments for publication. See \href{https://developer.datawrapper.de/docs/chart-properties-1}{the documentation} for details.}
}
\value{
A terminal message: "Chart xyz succesfully updated." - or an error message.
}
\description{
Modifies an existing Datawrapper chart.
}
\note{
This function builds a body for a API-call to the Datawrapper-API, which contains changes to an existing chart.

Check their \href{https://developer.datawrapper.de/docs/reference-guide}{reference guide for examples}.
}
\examples{

\dontrun{dw_edit_chart("aBcDE")} # uses the preset key in the .Renviron-file

\dontrun{dw_edit_chart(chart_id = "a1B2Cd", api_key = "1234ABCD")} # uses the specified key

\dontrun{dw_edit_chart(chart_id = "a1B2Cd", title = "I'm a title",
intro = "Data showing daily results")} # changes title and intro

\dontrun{dw_edit_chart(chart_id = "a1B2Cd", title = "I'm a title",
visualize = list(`show-tooltips` = "false"))} # adds manual changes as lists

}
\author{
Benedict Witzenberger
}
