% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dw_export_chart.R
\name{dw_export_chart}
\alias{dw_export_chart}
\title{Export a Datawrapper chart to PNG, PDF, or SVG}
\usage{
dw_export_chart(
  chart_id,
  type = c("png", "pdf", "svg"),
  unit = c("px", "inch", "mm"),
  mode = c("rgb", "cmyk"),
  width = NULL,
  height = NULL,
  plain = TRUE,
  scale = 2,
  border_width = 0,
  border_color = NULL,
  api_key = "environment"
)
}
\arguments{
\item{chart_id}{Required. A Datawrapper-chart-id as character string, usually a five character combination of digits and letters, e.g. "aBcDe". Or a \strong{dw_chart}-object.}

\item{type}{Required. One of \code{png}, \code{pdf}, or \code{svg}. Defaults to \code{png}.}

\item{unit}{Define units which \code{border_width}, \code{height} and \code{width} are measured in. \code{png} only works with \code{px};
\code{pdf} works with any of them; \code{svg} should have this set to \code{NULL}.}

\item{mode}{color movde for output. Defaults to \code{rgb}. Only applicable to \code{pdf} output type.}

\item{width, height}{width and height of the visualization; if not specified it takes the chart width/height.}

\item{plain}{if \code{TRUE} (the default) only the visualization will be exported; \code{FALSE} includes header and footer.}

\item{scale}{defines multiplier for the size; e.g. if \code{3} then the chart will be 3x wider and taller; Defaults to \code{2}.}

\item{border_width}{margin around the visualization; e.g. if \code{border_width} is \code{20} and \code{unit} is \code{px}, the visualization
will have a 20px margin. Default is \code{0}.}

\item{border_color}{color of the border; The default is the same as the visualization (likely \code{white}).}

\item{api_key}{Optional. A Datawrapper-API-key as character string. Defaults to "environment" - tries to automatically retrieve the key that's stored in the .Reviron-file by \code{\link{datawrapper_auth}}.}
}
\value{
for \code{png}, a \code{{magick}} object; for \code{pdf}, invisible raw vector with PDF content; for \code{svg}, invisible character vector with SVG content
}
\description{
Uses the \href{https://developer.datawrapper.de/docs/exporting-as-pdfsvg}{export endpoint} to export a chart as PNG, PDF, or SVG
(the latter two require paid plans).
}
\note{
PDF and SVG export require paid plans
}
\examples{
\dontrun{dw_export_chart("aBcDE")} # uses the preset key in the .Renviron-file
}
\author{
Bob Rudis \href{mailto:bob@rud.is}{bob@rud.is}
}
