% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dw_get_api_key.R
\name{dw_get_api_key}
\alias{dw_get_api_key}
\title{Get Datawrapper-API-Key}
\usage{
dw_get_api_key()
}
\value{
The API-key as a character string
}
\description{
Retrieves the \href{https://developer.datawrapper.de/}{Datawrapper-API}-key from the \code{DW_KEY}
environment variable if it exists.
}
\details{
You can use \code{\link[=datawrapper_auth]{datawrapper_auth()}} to store your API key into your local \code{.Renviron} file.
}
\note{
This is a helper function, that shouldn't be of any use outside of this package.
}
\examples{

\dontrun{dw_get_api_key()}

}
\author{
Benedict Witzenberger
}
