% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dw_get_map_key.R
\name{dw_get_map_key}
\alias{dw_get_map_key}
\title{Returns the Datawrapper-Map-keys}
\usage{
dw_get_map_key(api_key = "environment", basemap_id, basemap_value)
}
\arguments{
\item{api_key}{Required. A Datawrapper-API-key as character string. Defaults to "environment" - tries to automatically retrieve the key that's stored in the .Reviron-file by \code{\link{datawrapper_auth}}.}

\item{basemap_id}{Required. Can be retrieved by looking at \code{\link{dw_basemaps}}.}

\item{basemap_value}{Required. Can be retrieved by looking at \code{\link{dw_basemaps}}.}
}
\value{
A S3-structure of type \strong{dw_basemap_keys} with the elements from the \href{https://developer.datawrapper.de/}{Datawrapper-API}, and the following fields:
\item{content}{A data.frame of the returned keys.}
\item{path}{The url of the API call.}
\item{id}{The basemap id provided.}
\item{value}{The basemap value provided.}
}
\description{
Retrieve the keys for a certain basemap from the \href{https://developer.datawrapper.de/}{Datawrapper-API}.
}
\note{
This function returns a data.frame with the keys for the requested map. These keys can be used to map data to the basemap.
}
\examples{

\dontrun{

dw_get_map_key(basemap_id = "world2019", basemap_value = "DW_NAME") # uses the preset key in the .Renviron-file

}

}
\author{
Benedict Witzenberger
}
