% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dw_handle_errors.R
\name{dw_handle_errors}
\alias{dw_handle_errors}
\title{Handle errors in API call}
\usage{
dw_handle_errors(r)
}
\arguments{
\item{r}{Required. A httr-response-object.}
}
\value{
A list with the content of the respone-object - or an error,
}
\description{
Internal function: Handles errors while calling the Datawrapper-API.
}
\examples{

\dontrun{dw_handle_errors(r)}

}
\author{
Benedict Witzenberger
}
\keyword{Internal}
