% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dw_invite_user.R
\name{dw_invite_user}
\alias{dw_invite_user}
\title{Invite a person to a team}
\usage{
dw_invite_user(
  team,
  email,
  role = c("member", "admin", "owner"),
  api_key = "environment"
)
}
\arguments{
\item{team}{Required. The team-id (can be found in the URL of a teams-folder) as character.}

\item{email}{Required. The email-address that gets an invitation as character.}

\item{role}{Required. Set to one of: \code{member}, \code{admin}, \code{owner}. Check this for more informations: \href{https://academy.datawrapper.de/article/212-how-to-invite-others-to-a-team}{How to invite others to a team}}

\item{api_key}{Optional. A Datawrapper-API-key as character string. Defaults to "environment" - tries to automatically retrieve the key that's stored in the .Reviron-file by \code{\link{datawrapper_auth}}.}
}
\value{
A message of failure or success, derived from the API's status code.
}
\description{
\lifecycle{maturing}
}
\details{
Invites a new team member.
}
\note{
This function invites a email-address to a team.
}
\examples{

\dontrun{
dw_invite_user(team = "testTeam", email = "test_at_test.com", role = "admin")
} # uses the preset key in the .Renviron-file
}
\author{
Benedict Witzenberger
}
