% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dw_list_charts.R
\name{dw_list_charts}
\alias{dw_list_charts}
\title{Lists all charts}
\usage{
dw_list_charts(
  api_key = "environment",
  userId = "",
  published = "",
  search = "",
  order = "DESC",
  orderBy = "createdAt",
  limit = ""
)
}
\arguments{
\item{api_key}{Required. A Datawrapper-API-key as character string. Defaults to "environment" - tries to automatically retrieve the key that's stored in the .Reviron-file by \code{\link{datawrapper_auth}}.}

\item{userId}{Optional. A numeric. Only return charts created by a certain user id.}

\item{published}{Optional. A string. Use true or false (lowercase) to use this filter.}

\item{search}{Optional. A string. Use this filter to search for a certain character string.}

\item{order}{Optional. A string. DESC by default, might be changed to "ASC".}

\item{orderBy}{Optional. A string. Is set to "createdAt" by default. Might be changed to: "id", "email", "name".}

\item{limit}{Optional. A numeric. Defaults to 100. Number of charts to be retrieved.}
}
\value{
A tibble of all created charts. Not including all metadata, which can be retrievd for a single chart using \code{\link{dw_retrieve_chart_metadata}}.
\item{total}{Total number of created charts.}
\item{next}{Returns the API-call for the next page}
}
\description{
Returns all created Datawrapper charts by the current user.
}
\examples{

\dontrun{dw_list_charts("aBcDE")} # uses the preset key in the .Renviron-file

\dontrun{response_list <- dw_list_charts()} # save list of charts into variable

}
\author{
Benedict Witzenberger, Bob Rudis
}
