% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dw_list_folders.R
\name{dw_list_folders}
\alias{dw_list_folders}
\title{Lists all folders}
\usage{
dw_list_folders(api_key = "environment")
}
\arguments{
\item{api_key}{Optional. A Datawrapper-API-key as character string. Defaults to "environment" - tries to automatically retrieve the key that's stored in the .Reviron-file by \code{\link{datawrapper_auth}}.}
}
\value{
Returns a large list of all folders, created by user or team, which then includes all folders with their designated charts.
}
\description{
\lifecycle{maturing}
Returns all existing folders
}
\note{
This function builds a body for a API-call to the Datawrapper-API, and retrieves all folders a user or team have created.
}
\examples{

\dontrun{dw_list_folders()} # uses the preset key in the .Renviron-file

\dontrun{response_list <- dw_list_folders()} # save the result to a variable

}
\author{
Benedict Witzenberger
}
