% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dw_retrieve_chart_metadata.R
\name{dw_retrieve_chart_metadata}
\alias{dw_retrieve_chart_metadata}
\title{Retrieves a Datawrapper chart's metadata}
\usage{
dw_retrieve_chart_metadata(chart_id, api_key = "environment")
}
\arguments{
\item{chart_id}{Required. A Datawrapper-chart-id as character string, usually a five character combination of digits and letters, e.g. "aBcDe".}

\item{api_key}{Optional. A Datawrapper-API-key as character string. Defaults to "environment" - tries to automatically retrieve the key that's stored in the .Reviron-file by \code{\link{datawrapper_auth}}.}
}
\value{
A S3-structure of type dw_chart with the elements from the Datawrapper-API stored under content.
\item{status}{Returns 'ok' if the API-key used was correct.}
\item{$data$id}{Returns the internal id of the chart - the same as used in chart_id.}
\item{$data$title}{Returns the chart's title.}
\item{$data$theme}{Returns the chart's theme.}
\item{$data$createdAt}{Chart's creation date.}
\item{$data$lastModifiedAt}{Chart's last modification date.}
\item{$data$metadata}{Contains chart's specifications, like transpose, column formats, visualization settings, colors, titles and texts}
\item{$data$metadata$visualize}{Contains the chart's visualization settings.}
\item{$data$metadata$describe}{Contains the chart's description settings, like title, source name and url, byline}
\item{$data$publishedAt}{Chart's publication date - if published yet.}
\item{$data$author$id}{The chart-author's id.}
}
\description{
Return the metadata of a existing Datawrapper chart.
}
\note{
This function retrieves all metadata about a chart that's stored by Datawrapper. It is helpful to gain insights in the different options that might be changed via the API.
}
\examples{

\dontrun{dw_retrieve_chart_metadata("aBcDE")} # uses the preset key in the .Renviron-file

\dontrun{dw_retrieve_chart_metadata(chart_id = "a1B2Cd", api_key = "1234ABCD")} # uses the specified key
}
\author{
Benedict Witzenberger
}
