% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dw_test_key.R
\name{dw_test_key}
\alias{dw_test_key}
\title{Tests the Datawrapper-API-key}
\usage{
dw_test_key(api_key = "environment")
}
\arguments{
\item{api_key}{Required. A Datawrapper-API-key as character string. Defaults to "environment" - tries to automatically retrieve the key that's stored in the .Reviron-file by \code{\link{datawrapper_auth}}.}
}
\value{
A S3-structure of type \strong{dw_user} with the elements from the \href{https://developer.datawrapper.de/}{Datawrapper-API}, and the following fields:
\item{status}{Returns 'ok' if the API-key used was correct.}
\item{data$user$id}{Returns the internal user id.}
\item{data$user$email}{The users e-mail adress.}
\item{data$user$isAdmin}{Specifies, if the current user is admin in his/her organization.}
\item{data$user$organization$id}{Returns the organization's id or short name.}
\item{data$user$organization$name}{Returns the organization's full name.}
}
\description{
Tests the key by querying the \href{https://developer.datawrapper.de/}{Datawrapper-API} for user information.
}
\note{
This function tests the API key by retrieving information about the current user from the API. If this works, the API-key is set correctly and ready to go.
}
\examples{

\dontrun{dw_test_key()} # uses the preset key in the .Renviron-file

\dontrun{dw_test_key(api_key = "1234ABCD")} # uses the specified key

}
\author{
Benedict Witzenberger
}
