% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dw_create_chart.R
\name{dw_create_chart}
\alias{dw_create_chart}
\title{Creates a new Datawrapper chart}
\usage{
dw_create_chart(api_key = "environment", title = "", type = "")
}
\arguments{
\item{api_key}{Optional. A Datawrapper-API-key as character string. Defaults to "environment" - tries to automatically retrieve the key that's stored in the .Reviron-file by \code{\link{datawrapper_auth}}.}

\item{title}{Optional. Will set a chart's title on creation.}

\item{type}{Optional. Changes the type of the chart. See \href{https://developer.datawrapper.de/docs/chart-types-2}{the documentation} for the different types.}
}
\value{
It prints the new chart's id and returns a S3-structure of type \strong{dw_chart} with the elements from the Datawrapper-API, the same as in \code{\link{dw_retrieve_chart_metadata}}.
}
\description{
Creates and returns a new Datawrapper chart object.
}
\note{
If not specified, the new chart will by default be created without a title and with the type \code{d3-lines}.
}
\examples{

\dontrun{
dw_create_chart()
} # uses the preset key in the .Renviron-file

\dontrun{dw_create_chart(title = "Testtitle")}

\dontrun{dw_create_chart(api_key = "1234ABCD")} # uses the specified key
}
\author{
Benedict Witzenberger
}
