% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dw_legend_to_string.R
\name{dw_legend_to_string}
\alias{dw_legend_to_string}
\title{Creates a legend from a Datawrapper chart}
\usage{
dw_legend_to_string(
  legend,
  return_val = c("string", "vector", "html"),
  use_ranges = FALSE
)
}
\arguments{
\item{legend}{Required. A vector containing the legend (e.g. retrieved from a call with \code{\link{dw_retrieve_chart_metadata(chart_id)}}).}

\item{return_val}{Required. One of c("string", "vector", "html"). Set the output type. Defaults to "string".}

\item{use_ranges}{Optional. Filters out automatically generated ranges (e.g. \verb{(0,10]}). Defaults to FALSE.}
}
\value{
A string, vector or HTML (as string), as defined in \code{return_val}.
}
\description{
\lifecycle{deprecated}
A helper function that creates a legend as string, vector or HTML from an existing legend in a Datawrapper chart.
Will be deprecated soon, as it's no longer needed. Datawrapper now offers \href{https://blog.datawrapper.de/choropleth-symbol-maps-easier-faster-better-looking/}{a better option via its interface}.
}
\section{How to}{


Create a chart that displays the required legend. Then use \code{\link{dw_retrieve_chart_metadata(chart_id)}} to download the meta-information of your chart into a R-variable.
You may then extract the legend as a string. It might for example be stored in \code{metadata[["content"]][["metadata"]][["visualize"]][["categories"]]} or in \code{metadata[["content"]][["metadata"]][["visualize"]][["custom-colors"]]}.
}

\examples{

\dontrun{

dw_legend_to_string(legend)

dw_legend_to_string(legend, return_val = "html") # return simple HTML-code

dw_legend_to_string(legend, use_ranges = TRUE) # filter out ranges
}
}
\author{
Benedict Witzenberger
}
