% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dw_publish_chart.R
\name{dw_publish_chart}
\alias{dw_publish_chart}
\title{Publish or republish a Datawrapper chart}
\usage{
dw_publish_chart(chart_id, api_key = "environment", return_urls = TRUE)
}
\arguments{
\item{chart_id}{Required. A Datawrapper-chart-id as character string, usually a five character combination of digits and letters, e.g. "aBcDe". Or an \emph{dw_chart}-object.}

\item{api_key}{Optional. A Datawrapper-API-key as character string. Defaults to "environment" - tries to automatically retrieve the key that's stored in the .Reviron-file by \code{\link{datawrapper_auth}}.}

\item{return_urls}{Optional. If TRUE (default) it returns the code for the responsive iFrame and an URL to the chart.}
}
\value{
A message that specifies, if the publication was successful. If set, including the iFrame-Code and chart-URL.
}
\description{
Publish a chart on Datawrapper.
}
\note{
This function publishes a chart in Datawrapper.
}
\examples{

\dontrun{dw_publish_chart("aBcDE")} # uses the preset key in the .Renviron-file

\dontrun{dw_publish_chart(chart_id = "a1B2Cd", api_key = "1234ABCD")} # uses the specified key

\dontrun{dw_publish_chart(chart_id = "a1B2Cd", return_urls = TRUE)} # won't return code and URLs for chart

}
\author{
Benedict Witzenberger
}
